add dataclass transform
