+++ requirements.txt
@@ -14,3 +14,4 @@ autoray==0.3.1
 matplotlib~=3.5
 opt_einsum~=3.3
 requests~=2.28.1
+typing_extensions~=4.5.0
